+++ src/java.base/share/classes/jdk/internal/foreign/MappedMemorySegmentImpl.java
@@ -72,23 +72,28 @@ final class MappedMemorySegmentImpl extends NativeMemorySegmentImpl {
 
     public void load() {
         if (unmapper != null) {
-            SCOPED_MEMORY_ACCESS.load(sessionImpl(), min, unmapper.isSync(), length);
+            SCOPED_MEMORY_ACCESS.load(sessionImpl(), NIO_ACCESS.mappedMemoryUtils(),
+                    min, unmapper.isSync(), length);
         }
     }
 
     public void unload() {
         if (unmapper != null) {
-            SCOPED_MEMORY_ACCESS.unload(sessionImpl(), min, unmapper.isSync(), length);
+            SCOPED_MEMORY_ACCESS.unload(sessionImpl(), NIO_ACCESS.mappedMemoryUtils(),
+                    min, unmapper.isSync(), length);
         }
     }
 
     public boolean isLoaded() {
-        return unmapper == null || SCOPED_MEMORY_ACCESS.isLoaded(sessionImpl(), min, unmapper.isSync(), length);
+        return unmapper == null ||
+                SCOPED_MEMORY_ACCESS.isLoaded(sessionImpl(),
+                        NIO_ACCESS.mappedMemoryUtils(), min, unmapper.isSync(), length);
     }
 
     public void force() {
         if (unmapper != null) {
-            SCOPED_MEMORY_ACCESS.force(sessionImpl(), unmapper.fileDescriptor(), min, unmapper.isSync(), 0, length);
+            SCOPED_MEMORY_ACCESS.force(sessionImpl(), NIO_ACCESS.mappedMemoryUtils(),
+                    unmapper.fileDescriptor(), min, unmapper.isSync(), 0, length);
         }
     }
